mnk/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.1.0
sqlparse==0.5.3
tzdata==2025.1
Pillow==11.1.0
psycopg2==2.9.10
django-oauth-toolkit==3.0.1
oauthlib==3.2.2